Struct gp_resonator::resonator_array::ConjPoleResonatorArray

pub struct ConjPoleResonatorArray {
    pub sample_rate: f64,
    /* private fields */
}

A resonator composed of a dynamic number of ConjPoleResonators. See ConjPoleResonatorArray::process_buf.

Fields
sample_rate: f64

The sample rate of audio being passed through the resonator

Implementations

impl ConjPoleResonatorArray

pub fn new(sample_rate: f64, capacity: usize) -> Self

Creates a new instance of ConjPoleResonatorArray with given sample rate and capacity. The capacity passed is only used to allocating the initial memory for the resonator array and is mainly for efficiency (any value can be passed).

Arguments

sample_rate - the sample rate of the audio being passed to the system (used to determine pole locations given a frequency in Hz)

capacity - a soft capacity size for the resonator array

pub fn process_single(&self, x_1: f64, y_1: f64, y_2: f64) -> f64

Processes a single data point given the required values from the difference equation and returns y[n]. Prefer Self::process_buf for better cache performance.

Arguments

x_1 - value of x[n-1]

y_1 - value of y[n-1]

y_2 - value of y[n-2]

pub fn process_buf(&mut self, x: &[f64], buf: &mut [f64])

Takes a signal and adds the output of the signal going through the resonator array to a given buffer.

Calling this updates the resonator array so that the next call will be treated as if the given signal is concatenated to the previous signal. To prevent this, one may call Self::reset_state.

Arguments

x - signal buffer

buf - buffer to write to

pub fn reset_state(&mut self)

Resets the state of the internal resonators. The state of the internal resonators is used to allow continuous signal processing in given chunks. See Self::process_buf.

pub fn add_resonator(
&mut self,
freq: f64,
intensity: f64,
gain: f64
) -> Result<(), &'static str>

Add a resonator with given intensity and gain to the array.

Arguments

freq - the frequency of the resonator in Hz

intensity - the modulus of the pole in the complex plane (must be in range [0, 1) for stable system)

gain - the gain of the resonator output (amplitude)